refactored api token functionality into separate file
