+++ apps/frontend/lib/api/token.ts
@@ -1,0 +1,27 @@
+import { fromApiEndpoint, getResource } from "./common"
+
+interface TokenBase {
+  ticker: string
+  name: string
+  evm_contract_address: string
+  abi: any
+}
+
+interface Token extends TokenBase {
+  id: string
+}
+
+async function getToken(tokenId: string): Promise<Token> {
+  return getResource<Token>(
+    fromApiEndpoint('tokens/'),
+    { resourceId: tokenId },
+  )
+}
+
+export {
+  getToken,
+}
+
+export type {
+  Token,
+}
